move response-tooltip's relations functions to dedicated file
